added route response

+++ location.go
@@ -1,7 +1,7 @@
 package valhalla
 
 type Location struct {
-	Lat    float32 `json:"lat,omitempty"`
-	Lon    float32 `json:"lon,omitempty"`
+	Lat    float64 `json:"lat,omitempty"`
+	Lon    float64 `json:"lon,omitempty"`
 	Street string  `json:"street,omitempty"`
 }
